hass: Major revamp of config
This is a major revamp of my entire Home Assistant config and there are
too many things to list to that has changed, but a few to mention:

  * Tellstick Net replaced by rflink
  * Projector and projector screen added
  * New custom component using CEC for TV (mostly a hack)
  * Migrated to Lovelace (everything related to "States" removed)
  * Evertyhing I don't use have been removed
  * Deprecated components are under homeassistant/deprecated/
  * Custom theme based on "darkblue"
  * Using LIRC + IR receiver in order to hide my ATV 4k in TV bench
    (allows me to use volume controls on remote)
  * Migrated garage sensor to esphomelib

Various minor changes as well, e.g. using filesize sensor for monitoring
log size and so on.

My intentions are to keep my setup as clean and minimalistic as possible
to simplify management. I want it to run smoothly. Later on I will move
to use the built-in editor for Lovelace as well, currently using
YAML-mode.